requirements.txt
numpy 
pandas
tensorflow
matplotlib
seaborn
nltk
regex
scikit-learn
from-root
google-cloud-storage
pyaml
# -e .